refactor: replace 87 fragmented E2E tests with 15 journey tests
Delete 8 authenticated test files (87 tests) that mostly checked
element visibility, and replace with 7 journey test files (15 tests)
that click through complete user flows.

Coverage preserved:
- Task list, grouping, detail, verdict buttons, previous step output
- Workflow home cards, run counts, navigation to run detail
- Run detail step graph, step history, audit log, autonomy badges
- Cancel run flow with confirm/dismiss
- Run report timeline, detail toggle, branding
- Workflow editor tabs, definitions, diagram, edit mode
- Agent catalog, run history, agent detail, escalated runs

Coverage intentionally dropped:
- CSS class assertions (border-l-4, opacity-60, font-medium)
- Duplicate route-exists smoke tests
- Navigation spec (absorbed into journeys)

Result: 18 total tests (was 87), 34s runtime (was 45s), all passing.
