--- packages/platform-ui/e2e/authenticated/run-report.spec.ts
@@ -1,70 +1,0 @@
-import { test, expect } from '@playwright/test';
-import { TEST_ORG_HANDLE } from '../helpers/constants';
-
-// Uses seeded data from auth-setup.ts:
-// - 'proc-completed-1' is a completed instance with stepExecutions subcollection seeded
-// - definitionName: 'Data Quality Review', definitionVersion: '2.1.0'
-// - stepExecutions: 'exec-dq-verify' (completed), 'exec-dq-review' (completed)
-
-test.describe('Run Report', () => {
-  const reportUrl = `/${TEST_ORG_HANDLE}/workflows/Data%20Quality%20Review/runs/proc-completed-1/report`;
-  const runUrl = `/${TEST_ORG_HANDLE}/workflows/Data%20Quality%20Review/runs/proc-completed-1`;
-
-  test('[RENDER] report page loads for completed run', async ({ page }) => {
-    await page.goto(reportUrl);
-    await expect(page.getByRole('heading', { name: /Data Quality Review — Run Report/i })).toBeVisible();
-  });
-
-  test('[RENDER] report shows step timeline with step names', async ({ page }) => {
-    await page.goto(reportUrl);
-    await expect(page.getByText(/Step Timeline/i)).toBeVisible();
-    // Step names from seed data or formatted step IDs
-    await expect(page.getByText(/Verify Data Quality|verify-data-quality/i).first()).toBeVisible();
-    await expect(page.getByText(/Review Results|review-results/i).first()).toBeVisible();
-  });
-
-  test('[RENDER] report shows timing information', async ({ page }) => {
-    await page.goto(reportUrl);
-    await expect(page.getByText(/Wall-clock/i)).toBeVisible();
-    await expect(page.getByText(/Active processing/i)).toBeVisible();
-  });
-
-  test('[CLICK] detail level toggle switches between brief and full', async ({ page }) => {
-    await page.goto(reportUrl);
-
-    // Default is brief — should show executor info but no audit trail
-    await expect(page.getByText(/Executed by/i).first()).toBeVisible();
-
-    // Switch to full — audit trail and input/output should appear
-    await page.getByRole('button', { name: /full/i }).click();
-    await expect(page.getByText(/Output/i).first()).toBeVisible();
-
-    // Switch back to brief
-    await page.getByRole('button', { name: /brief/i }).click();
-    await expect(page.getByText(/Step Timeline/i)).toBeVisible();
-  });
-
-  test('[RENDER] report has logo and footer branding', async ({ page }) => {
-    await page.goto(reportUrl);
-    await expect(page.getByAltText('Mediforce')).toBeVisible();
-    await expect(page.getByText(/Generated by Mediforce/i)).toBeVisible();
-  });
-
-  test('[RENDER] print button is present', async ({ page }) => {
-    await page.goto(reportUrl);
-    await expect(page.getByRole('button', { name: /print/i })).toBeVisible();
-  });
-
-  test('[RENDER] View Report link on completed run page', async ({ page }) => {
-    await page.goto(runUrl);
-    const reportLink = page.getByRole('link', { name: /View Report/i });
-    await expect(reportLink).toBeVisible();
-    await expect(reportLink).toHaveAttribute('href', /\/report$/);
-  });
-
-  test('[RENDER] report shows not available for non-completed runs', async ({ page }) => {
-    // proc-running-1 is a running instance
-    await page.goto(`/${TEST_ORG_HANDLE}/workflows/Supply%20Chain%20Review/runs/proc-running-1/report`);
-    await expect(page.getByText(/only available for completed runs/i)).toBeVisible();
-  });
-});
