+++ packages/platform-ui/e2e/journeys/run-report.journey.ts
@@ -1,0 +1,46 @@
+import { test, expect } from '@playwright/test';
+import { TEST_ORG_HANDLE } from '../helpers/constants';
+
+test.describe('Run Report Journey', () => {
+  test('completed run report shows timeline, toggles detail level, and has branding', async ({ page }) => {
+    // First check the View Report link exists on run detail
+    const runUrl = `/${TEST_ORG_HANDLE}/workflows/Data%20Quality%20Review/runs/proc-completed-1`;
+    await page.goto(runUrl);
+    const reportLink = page.getByRole('link', { name: /View Report/i });
+    await expect(reportLink).toBeVisible({ timeout: 10_000 });
+    await expect(reportLink).toHaveAttribute('href', /\/report$/);
+
+    // Navigate to report
+    const reportUrl = `${runUrl}/report`;
+    await page.goto(reportUrl);
+    await expect(page.getByRole('heading', { name: /Data Quality Review — Run Report/i })).toBeVisible({ timeout: 10_000 });
+
+    // Step Timeline and step names
+    await expect(page.getByText(/Step Timeline/i)).toBeVisible();
+    await expect(page.getByText(/Verify Data Quality|verify-data-quality/i).first()).toBeVisible();
+    await expect(page.getByText(/Review Results|review-results/i).first()).toBeVisible();
+
+    // Timing information
+    await expect(page.getByText(/Wall-clock/i)).toBeVisible();
+    await expect(page.getByText(/Active processing/i)).toBeVisible();
+
+    // Toggle detail level: brief -> full -> brief
+    await expect(page.getByText(/Executed by/i).first()).toBeVisible();
+    await page.getByRole('button', { name: /full/i }).click();
+    await expect(page.getByText(/Output/i).first()).toBeVisible();
+    await page.getByRole('button', { name: /brief/i }).click();
+    await expect(page.getByText(/Step Timeline/i)).toBeVisible();
+
+    // Branding
+    await expect(page.getByRole('img', { name: 'Mediforce', exact: true })).toBeVisible();
+    await expect(page.getByText(/Generated by Mediforce/i)).toBeVisible();
+
+    // Print button
+    await expect(page.getByRole('button', { name: /print/i })).toBeVisible();
+  });
+
+  test('report unavailable for non-completed runs', async ({ page }) => {
+    await page.goto(`/${TEST_ORG_HANDLE}/workflows/Supply%20Chain%20Review/runs/proc-running-1/report`);
+    await expect(page.getByText(/only available for completed runs/i)).toBeVisible({ timeout: 10_000 });
+  });
+});
