+++ packages/platform-ui/e2e/smoke.spec.ts
@@ -13,17 +13,4 @@ test.describe('Smoke Tests', () => {
     await page.waitForURL('**/login**');
     await expect(page).toHaveURL(/\/login/);
   });
-
-  test('[RENDER] /{handle} profile page route exists and responds', async ({ page }) => {
-    const response = await page.goto('/unknown-test-handle-xyz');
-    // Route must exist — any HTTP response (200, 404, 500) confirms the route is registered.
-    // A 500 in CI without Firebase credentials is expected and acceptable here.
-    expect(response?.status()).toBeGreaterThanOrEqual(200);
-  });
-
-  test('[RENDER] /{handle} profile+workflows page route exists and responds', async ({ page }) => {
-    const response = await page.goto('/unknown-test-handle-xyz');
-    // Route must exist — any HTTP response confirms routing works.
-    expect(response?.status()).toBeGreaterThanOrEqual(200);
-  });
 });
