+++ packages/platform-ui/package.json
@@ -15,6 +15,8 @@
     "test:e2e:auth:headed": "NEXT_PUBLIC_USE_EMULATORS=true playwright test --headed",
     "test:e2e:ui": "playwright test --ui",
     "test:e2e:headed": "playwright test --headed",
+    "test:e2e:record": "NEXT_PUBLIC_USE_EMULATORS=true E2E_RECORD=true playwright test --project=authenticated --workers=1",
+    "test:e2e:gif": "NEXT_PUBLIC_USE_EMULATORS=true E2E_RECORD=true playwright test --project=authenticated --workers=1 && echo 'Convert videos: find test-results -name *.webm -exec sh -c '\"'\"'ffmpeg -i {} -vf fps=10,scale=960:-1:flags=lanczos -loop 0 {}.gif'\"'\"' \\;'",
     "emulators": "bash scripts/start-emulators.sh",
     "deploy": "pnpm build && firebase deploy --only hosting"
   },
